% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smartsheet-R-sdk.R
\name{cell_history}
\alias{cell_history}
\title{Get the history of a cell}
\usage{
cell_history(sheet_name, rowId, columnId)
}
\arguments{
\item{sheet_name}{The name of the sheet with your cell}

\item{rowId}{The ID of the row with your cell. You can find this by
right-clicking the row and selecting Properties}

\item{columnId}{The ID of the column with your row. Can be found
with [get_sheet_column_info()].}
}
\value{
returns a tibble with the cell's history
}
\description{
Get the history of a cell
}
