% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smartsheet-R-sdk.R
\name{check_smartsheet_api_key}
\alias{check_smartsheet_api_key}
\title{Check Smartsheet API Key}
\usage{
check_smartsheet_api_key()
}
\value{
if this function works it returns nothing
}
\description{
Check Smartsheet API Key
}
\examples{
\dontrun{
check_smartsheet_api_key()
}
}
