% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smartsheet-R-sdk.R
\name{csv_to_sheet}
\alias{csv_to_sheet}
\title{Create New Smartsheet}
\usage{
csv_to_sheet(file_path, all_text_number = FALSE)
}
\arguments{
\item{file_path}{a path which locates the csv and provides the name}
}
\value{
returns nothing
}
\description{
Create New Smartsheet
}
\examples{
\dontrun{
csv_to_sheet("a_folder/maybe_another_folder/sheet_name.csv", 123456789)
csv_to_sheet("a_folder/maybe_another_folder/sheet_name.csv", "123456789")
}
}
