% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smartsheet-R-sdk.R
\name{csv_to_sheet_in_folder}
\alias{csv_to_sheet_in_folder}
\title{Create New Smartsheet in Folder with Specific ID}
\usage{
csv_to_sheet_in_folder(file_path, folder_id, all_text_number = FALSE)
}
\arguments{
\item{file_path}{a path which locates the csv and provides the name}

\item{folder_id}{a string or numeric that exactly matches the target folder ID}
}
\value{
returns nothing
}
\description{
Create New Smartsheet in Folder with Specific ID
}
\examples{
\dontrun{
csv_to_sheet_in_folder("a_folder/maybe_another_folder/sheet_name.csv", 123456789)
csv_to_sheet_in_folder("a_folder/maybe_another_folder/sheet_name.csv", "123456789")
}
}
