% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smartsheet-R-sdk.R
\name{delete_sheet_by_name}
\alias{delete_sheet_by_name}
\title{Delete an Existing Smartsheet in Folder with Specific ID}
\usage{
delete_sheet_by_name(sheet_name)
}
\arguments{
\item{sheet_name}{the target Smartsheets exact sheet name}
}
\value{
returns nothing
}
\description{
Delete an Existing Smartsheet in Folder with Specific ID
}
\examples{
\dontrun{
delete_sheet_by_name("sheet_name")
}
}
