% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smartsheet-R-sdk.R
\name{download_sheet_attachment}
\alias{download_sheet_attachment}
\title{Download a file of an existing attachment on a Smartsheet with a specific sheet name}
\usage{
download_sheet_attachment(sheet_name, file_path, attachments_name = "")
}
\arguments{
\item{sheet_name}{the target Smartsheet's exact sheet name}

\item{file_path}{the download path}

\item{attachments_name}{if the downlad path file name is not the name of the target attached file}
}
\value{
returns nothing but downloads file to location
}
\description{
Download a file of an existing attachment on a Smartsheet with a specific sheet name
}
\examples{
\dontrun{
download_sheet_attachment("sheet_name","a_folder/maybe_another_folder/sheet_name.csv")
download_sheet_attachment("sheet_name","a_folder/maybe_another_folder/sheet_name.csv", file_name="different_sheet_name")
}
}
