% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smartsheet-R-sdk.R
\name{get_sheet_as_csv}
\alias{get_sheet_as_csv}
\title{Get text CSV for an existing Smartsheet with a specific sheet name}
\usage{
get_sheet_as_csv(sheet_name)
}
\arguments{
\item{sheet_name}{the target Smartsheet's exact sheet name}
}
\value{
returns text of csv
}
\description{
Get text CSV for an existing Smartsheet with a specific sheet name
}
\examples{
\dontrun{
get_sheet_as_csv("sheet_name")
}
}
