% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smartsheet-R-sdk.R
\name{get_sheet_column_info}
\alias{get_sheet_column_info}
\title{Get Sheet Column Info}
\usage{
get_sheet_column_info(sheet_name)
}
\arguments{
\item{sheet_name}{name of smartsheet}
}
\value{
returns table of column info
}
\description{
Get Sheet Column Info
}
\examples{
\dontrun{
get_sheet_column_info("sheet_name")
}
}
