% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smartsheet-R-sdk.R
\name{get_sheet_csv_attachment}
\alias{get_sheet_csv_attachment}
\title{Get text CSV for an existing attachment on a Smartsheet with a specific sheet name}
\usage{
get_sheet_csv_attachment(sheet_name, attachments_name)
}
\arguments{
\item{sheet_name}{the target Smartsheet's exact sheet name}

\item{attachments_name}{the target attachment exact file name}
}
\value{
returns text of csv
}
\description{
Get text CSV for an existing attachment on a Smartsheet with a specific sheet name
}
\examples{
\dontrun{
get_sheet_as_csv("sheet_name", "atttachment.csv")
}
}
