% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smartsheet-R-sdk.R
\name{get_smart_folder_id}
\alias{get_smart_folder_id}
\title{Get Smartsheet Folder ID}
\usage{
get_smart_folder_id(folder_name)
}
\arguments{
\item{folder_name}{a string that exactly matches the target folder name}
}
\value{
the target folders ID as a string
}
\description{
Get Smartsheet Folder ID
}
\examples{
\dontrun{
get_smart_folder_id("Folder_Name")
}
}
