% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smartsheet-R-sdk.R
\name{list_reports}
\alias{list_reports}
\title{Get a list of reports that the user is able to access}
\usage{
list_reports()
}
\value{
A tibble of reports that the user is able to access.
}
\description{
Get a list of reports that the user is able to access
}
\examples{
\dontrun{
list_reports
}
}
