% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smartsheet-R-sdk.R
\name{list_workspaces}
\alias{list_workspaces}
\title{Get a list of workspaces that the user is able to access}
\usage{
list_workspaces()
}
\value{
A tibble of worksheets that the user is able to access.
}
\description{
Get a list of workspaces that the user is able to access
}
\examples{
\dontrun{
list_workbooks()
}
}
