% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smartsheet-R-sdk.R
\name{replace_sheet_attachment}
\alias{replace_sheet_attachment}
\title{Replace an existing attachment file on a Smartsheet with a new file}
\usage{
replace_sheet_attachment(sheet_name, file_path, attachment_name)
}
\arguments{
\item{sheet_name}{the target Smartsheet's exact sheet name}

\item{file_path}{the path to the file that will overwrite the existing attachment}

\item{attachment_name}{the name of the attached file to replace}
}
\value{
returns nothing
}
\description{
Replace an existing attachment file on a Smartsheet with a new file
}
\examples{
\dontrun{
replace_sheet_attachment("sheet_name","a_folder/maybe_another_folder/attachment_name.csv", "other_attachment_name.csv")
}
}
