% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smartsheet-R-sdk.R
\name{replace_sheet_with_csv}
\alias{replace_sheet_with_csv}
\title{Replace an existing Smartsheet with a new file}
\usage{
replace_sheet_with_csv(
  sheet_name,
  file_path,
  never_delete = FALSE,
  batch_size = 5000
)
}
\arguments{
\item{sheet_name}{the target Smartsheet's exact sheet name}

\item{file_path}{the path to the file that will overwrite the existing attachment}

\item{never_delete}{prevents deleting extra rows if the replacement csv has fewer rows than target sheet}
}
\value{
returns nothing
}
\description{
Replaces the data in all existing rows and then adds new rows as appropriate.
Throws errors if the same columns are not present in both the the Sheet and csv to repalce it.
}
\examples{
\dontrun{
replace_sheet_with_csv("sheet_name","a_folder/maybe_another_folder/sheet.csv")
}
}
