% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smartsheet-R-sdk.R
\name{set_sheet_columns_to_textnumber}
\alias{set_sheet_columns_to_textnumber}
\title{Set Sheet Columns to Text Number Types}
\usage{
set_sheet_columns_to_textnumber(sheet_name)
}
\arguments{
\item{sheet_name}{name of smartsheet}
}
\value{
returns nothing
}
\description{
Set Sheet Columns to Text Number Types
}
\examples{
\dontrun{
set_sheet_columns_to_textnumber("sheet_name")
}
}
