% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smartsheet-R-sdk.R
\name{set_smart_working_folder}
\alias{set_smart_working_folder}
\title{Set a working folder in Smartsheet with a folder ID}
\usage{
set_smart_working_folder(folder_id)
}
\arguments{
\item{folder_id}{a string or numeric that exactly matches the target folder ID}
}
\value{
if this function works it returns nothing
}
\description{
Set a working folder in Smartsheet with a folder ID
}
\examples{
\dontrun{
set_smart_working_folder("123456789")
set_smart_working_folder(123456789)
}
}
