% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smartsheet-R-sdk.R
\name{set_smartsheet_api_key}
\alias{set_smartsheet_api_key}
\title{Set the api key for acessing Smartsheet}
\usage{
set_smartsheet_api_key(key)
}
\arguments{
\item{key}{a string for the API key you want to use}
}
\value{
if this function works it returns nothing
}
\description{
Set the api key for acessing Smartsheet
}
\examples{
\dontrun{
set_smartsheet_api_key("yourAPIkey")
}
}
