% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smartsheet-R-sdk.R
\name{sheet_name_to_id}
\alias{sheet_name_to_id}
\title{Get Smartsheet Name From ID}
\usage{
sheet_name_to_id(sheet_name)
}
\arguments{
\item{sheet_name}{the name of a smartsheet
s}
}
\value{
returns sheet ID
}
\description{
Get Smartsheet Name From ID
}
\examples{
\dontrun{
sheet_name_to_id("sheet_name")
}
}
