% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smartsheet-R-sdk.R
\name{use_direct_ids}
\alias{use_direct_ids}
\title{Set whether you can use direct sheet id}
\usage{
use_direct_ids(inp)
}
\arguments{
\item{inp}{logical TRUE or FALSE}
}
\value{
returns nothing
}
\description{
Set whether you can use direct sheet id references instead of sheet names
}
\examples{
\dontrun{
use_direct_ids()
}
}
